Refactor headers obfuscation
